finish lab1

+++ src/lab1/include/print.h
@@ -1,7 +1,9 @@
 #ifndef _PRINT_H
 #define _PRINT_H
 
+#include "types.h"
+
 void puts(char *s);
-void puti(int x);
+void puti(uint64 x);
 
 #endif
